refactor leaderboard for extensibility
